Bug fixes
Clicking 0 would add an unknown value (temp1) to input as code was
copied from 1-9 where temp1 would be a value from 1-9.
Division by 0 would result in an endless loop. As division works as a
counting loop subtracting the divisor from the dividend until the
dividend is less than the divisor. When the divisor is 0, it constantly
subtracts 0.
Both of these issues have been fixed.